refactor(rules): extract RelaxedMockUsedForValueClassRule (paired)
testing_quality.go and registry_testing_quality.go follow a split
pattern: rule struct + helpers live in the former, the Check closure
lives in the latter. Demonstrate the paired-extraction shape by
moving everything for RelaxedMockUsedForValueClassRule into one new
file testing_quality_relaxed_mock.go:

- struct + Confidence()        (was testing_quality.go)
- primitiveTypes var           (was testing_quality.go)
- testingQualityTypeArgument   (was testing_quality.go)
- registerTestingQualityRelaxedMockUsedForValueClass (was registry_testing_quality.go)

All four artifacts were only referenced by this rule. The registry
caller registerTestingQualityRelaxedMockUsedForValueClass() still
fires from registerTestingQualityRules in registry_testing_quality.go.
Same Go package — no other call sites change.

testing_quality.go: 1983 → 1948 LOC.
registry_testing_quality.go: shrinks by 27 LOC.
